Mostly delete useWallet composable
